Solved convert to 24 with go
